♻️ Refactor[#6]: 경로 수정에 따른 전체 코드 수정
