+++ src/main/java/com/saeparam/HeyRoutine/global/error/handler/TokenHandler.java
@@ -1,8 +1,8 @@
-package com.saeparam.HeyRoutine.exception.handler;
+package com.saeparam.HeyRoutine.global.error.handler;
 
 
-import com.saeparam.HeyRoutine.exception.GeneralException;
-import com.saeparam.HeyRoutine.response.code.BaseErrorCode;
+import com.saeparam.HeyRoutine.global.error.exception.GeneralException;
+import com.saeparam.HeyRoutine.global.web.response.code.BaseErrorCode;
 
 public class TokenHandler extends GeneralException {
   public TokenHandler(BaseErrorCode errorCode) {
