+++ src/main/java/com/saeparam/HeyRoutine/global/web/response/code/BaseErrorCode.java
@@ -1,4 +1,4 @@
-package com.saeparam.HeyRoutine.response.code;
+package com.saeparam.HeyRoutine.global.web.response.code;
 
 public interface BaseErrorCode {
