+++ src/main/java/com/saeparam/HeyRoutine/global/web/response/code/status/ErrorStatus.java
@@ -1,8 +1,8 @@
-package com.saeparam.HeyRoutine.response.code.status;
+package com.saeparam.HeyRoutine.global.web.response.code.status;
 
 
-import com.saeparam.HeyRoutine.response.code.BaseErrorCode;
-import com.saeparam.HeyRoutine.response.code.ErrorReasonDTO;
+import com.saeparam.HeyRoutine.global.web.response.code.BaseErrorCode;
+import com.saeparam.HeyRoutine.global.web.response.code.ErrorReasonDTO;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
